Update cesm2git to deal with cime changes.
Add another hidden file that needs to be removed before updating clm
from svn. The subtree structure for cime/cime_config/cesm/allactive
changed with clm4_5_14_r229. Fixing the python to deal with the change
would be very complex. Since the subtree structure is going to change
again with the next cime update, it is not worth dealing with at this
time. The all active subtree in r229 was manually added.

Testing: imported r226-r228